Address some feedback from @rivernate
