+++ app/assets/javascripts/application.js
@@ -15,19 +15,5 @@
 // https://github.com/rails/sprockets#sprockets-directives) for details
 // about supported directives.
 //
-//= require jquery
-//= require jquery_ujs
 //= require turbolinks
-//= require bootstrap/affix
-//= require bootstrap/alert
-//= require bootstrap/button
-//= require bootstrap/carousel
-//= require bootstrap/collapse
-//= require bootstrap/dropdown
-//= require bootstrap/modal
-//= require bootstrap/tooltip
-//= require bootstrap/popover
-//= require bootstrap/scrollspy
-//= require bootstrap/tab
-//= require bootstrap/transition
 //= require_tree .
