+++ app/controllers/registration_controller.rb
@@ -7,16 +7,15 @@ class RegistrationController < ApplicationController
 
   def register
     # Fetch the a custom tool consumer profile
-    tcp_helper = Registration::ToolConsumerProfile.new(URI.parse(params[:tc_profile_url]), jwt_access_token)
+    tcp_helper = ToolConsumerProfile.new(URI.parse(params[:tc_profile_url]), jwt_access_token)
 
     # Do failure redirect to tool consumer if missing required
-    # TODO check for require services
-    redirect_to registration_failure_url and return unless tcp_helper.supports_required_capabilities?
+    redirect_to registration_failure_url('Missing required capabilities') and return unless tcp_helper.supports_required_capabilities?
 
     # Create a new tool proxy
     tool_proxy = ToolProxy.new(tcp_url: tcp_helper.tcp_url, base_url: request.base_url)
     tp_response = register_tool_proxy(tool_proxy, tcp_helper.tp_service_url, jwt_access_token)
-    tp_response_helper = Registration::ToolProxyResponse.new(tp_response, registration_redirect_url)
+    tp_response_helper = ToolProxyResponse.new(tp_response, registration_redirect_url)
 
     # Persist tool proxy and redirect if no errors
     if tp_response_helper.success?
@@ -24,7 +23,7 @@ class RegistrationController < ApplicationController
       tool_proxy.update_attributes(guid: tp_response_helper.tool_proxy_guid, shared_secret: shared_secret)
       redirect_to tp_response_helper.registration_success_url
     else
-      redirect_to registration_failure_url
+      redirect_to registration_failure_url('Error received from tool consumer')
     end
   end
 end
