+++ app/helpers/application_helper.rb
@@ -1,31 +1,2 @@
 module ApplicationHelper
-  def title(page_title)
-    content_for(:title) { page_title }
-  end
-
-  def meta_description(page_meta_description)
-    content_for(:meta_description) { page_meta_description }
-  end
-
-  def heading(page_heading)
-    content_for(:heading) { page_heading }
-  end
-
-  def humanize_boolean(input)
-    input ||= ''
-    case input.to_s.downcase
-    when 't', 'true'
-      'Yes'
-    else
-      'No'
-    end
-  end
-
-  def css_for_boolean(input)
-    if input
-      'success'
-    else
-      'danger'
-    end
-  end
 end
