+++ config/application.rb
@@ -46,9 +46,6 @@ module SimilarityDetectionReferenceTool
 
     # Set Redis as the back-end for the cache.
     config.cache_store = :redis_store, ENV['REDIS_CACHE_URL']
-
-    # Set Sidekiq as the back-end for Active Job.
-    config.active_job.queue_adapter = :sidekiq
     config.active_job.queue_name_prefix =
       "#{ENV['ACTIVE_JOB_QUEUE_PREFIX']}_#{Rails.env}"
