--- config/initializers/sidekiq.rb
@@ -1,9 +1,0 @@
-sidekiq_config = { url: ENV['ACTIVE_JOB_URL'] }
-
-Sidekiq.configure_server do |config|
-  config.redis = sidekiq_config
-end
-
-Sidekiq.configure_client do |config|
-  config.redis = sidekiq_config
-end
